hooks and components all are ready
